requirements.txt
requests
prettytable
2captcha-python
pysimplegui
svglib==1.0.1
python-decouple #this is to use .env file variables